\name{Gen}
\alias{Gen}
\alias{rs#}
\alias{chrom}
\alias{pos}
\title{
 Genotype data 
}
\description{
Numeric format of genotype dataset.
}
\usage{
  data(Gen)
}
\details{
Dataset input of Genotype for mrMLM function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
