\name{Genotype }
\alias{Genotype}
\alias{33-16}
\alias{Nov-38}
\alias{A4226}
\title{
 Genotype of real data 
}
\description{
Numeric format of genotype dataset.
}
\usage{
  data(Genotype)
}
\details{
Dataset input of Genotype for mrMLM function. 
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}
