\name{Phe}
\alias{Phe}
\alias{<phenotype>}
\alias{trait1}
\alias{trait2}
\alias{trait3}
\title{
 Phenotype dataset    
}
\description{
 Phenotype dataset of multiple traits. 
}
\usage{
  data(Phe)
}
\details{
 Dataset input of phenotype in mrMLM function.
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}
