\name{Phenotype}
\alias{Phenotype}
\alias{trait1}
\alias{trait2}
\alias{trait3}
\title{
 Phenotype of real data   
}
\description{
 Phenotype dataset of multiple traits. 
}
\usage{
  data(Phenotype)
}
\details{
 Dataset input of phenotype in mrMLM function.
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}
