\name{ResultFinal}
\alias{ResultFinal}
\alias{Chromosome}
\alias{QTN.effect}
\alias{LOD.score}
\title{
 Final result used to draw manhattan plot.
}
\description{
 Final result used to draw manhattan plot.
}
\usage{
 data(ResultFinal)
}
\details{
 Final result used to draw manhattan plot.
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}


