\name{ResultIntermediate}
\alias{ResultIntermediate}
\alias{Trait.ID}
\alias{Trait.name}
\alias{RS.}
\title{
 Intermediate result used to draw manhattan plot.
}
\description{
 Intermediate result used to draw manhattan plot.
}
\usage{
 data(ResultIntermediate)
}
\details{
 Intermediate result used to draw manhattan plot.
}
\author{
 Maintainer: Yuan-Ming Zhang<soyzhang@mail.hzau.edu.cn>
}

